move MODELS_BLACKLIST to lib
