+++ QgisModelBaker/processing_provider/ili2db_importing.py
@@ -32,11 +32,11 @@ from QgisModelBaker.libs.modelbaker.iliwrapper.ili2dbconfig import (
     UpdateDataConfiguration,
 )
 from QgisModelBaker.libs.modelbaker.iliwrapper.ili2dbutils import JavaNotFoundError
+from QgisModelBaker.libs.modelbaker.utils.globals import MODELS_BLACKLIST
 from QgisModelBaker.processing_provider.ili2db_algorithm import (
     Ili2gpkgAlgorithm,
     Ili2pgAlgorithm,
 )
-from QgisModelBaker.utils.gui_utils import MODELS_BLACKLIST
 
 
 class ProcessImporter(QObject):
@@ -157,6 +157,7 @@ class ProcessImporter(QObject):
     def get_configuration_from_input(self, parameters, context, tool):
 
         configuration = Ili2DbCommandConfiguration()
+        configuration.base_configuration = self.parent.current_baseconfig()
         configuration.tool = tool
 
         # get database settings form the parent
