+++ QgisModelBaker/utils/gui_utils.py
@@ -41,7 +41,7 @@ from qgis.PyQt.QtWidgets import (
 from qgis.PyQt.uic import loadUiType
 
 from QgisModelBaker.libs.modelbaker.iliwrapper.ilicache import IliCache
-from QgisModelBaker.libs.modelbaker.utils.globals import LogLevel
+from QgisModelBaker.libs.modelbaker.utils.globals import MODELS_BLACKLIST, LogLevel
 from QgisModelBaker.libs.modelbaker.utils.qt_utils import slugify
 from QgisModelBaker.utils.globals import CATALOGUE_DATASETNAME
 
@@ -67,75 +67,6 @@ TransferExtensions = [
     "XLS",
     "xlsx",
     "XLSX",
-]
-
-MODELS_BLACKLIST = [
-    "CHBaseEx_MapCatalogue_V1",
-    "CHBaseEx_WaterNet_V1",
-    "CHBaseEx_Sewage_V1",
-    "CHAdminCodes_V1",
-    "AdministrativeUnits_V1",
-    "AdministrativeUnitsCH_V1",
-    "WithOneState_V1",
-    "WithLatestModification_V1",
-    "WithModificationObjects_V1",
-    "GraphicCHLV03_V1",
-    "GraphicCHLV95_V1",
-    "NonVector_Base_V2",
-    "NonVector_Base_V3",
-    "NonVector_Base_LV03_V3_1",
-    "NonVector_Base_LV95_V3_1",
-    "GeometryCHLV03_V1",
-    "GeometryCHLV95_V1",
-    "Geometry_V1",
-    "InternationalCodes_V1",
-    "Localisation_V1",
-    "LocalisationCH_V1",
-    "Dictionaries_V1",
-    "DictionariesCH_V1",
-    "CatalogueObjects_V1",
-    "CatalogueObjectTrees_V1",
-    "AbstractSymbology",
-    "CodeISO",
-    "CoordSys",
-    "GM03_2_1Comprehensive",
-    "GM03_2_1Core",
-    "GM03_2Comprehensive",
-    "GM03_2Core",
-    "GM03Comprehensive",
-    "GM03Core",
-    "IliRepository09",
-    "IliSite09",
-    "IlisMeta07",
-    "IliVErrors",
-    "INTERLIS_ext",
-    "RoadsExdm2ben",
-    "RoadsExdm2ben_10",
-    "RoadsExgm2ien",
-    "RoadsExgm2ien_10",
-    "StandardSymbology",
-    "StandardSymbology",
-    "Time",
-    "Units",
-    "",
-    "CHAdminCodes_V2",
-    "AdministrativeUnits_V2",
-    "AdministrativeUnitsCH_V2",
-    "WithOneState_V2",
-    "WithLatestModification_V2",
-    "WithModificationObjects_V2",
-    "GraphicCHLV03_V2",
-    "GraphicCHLV95_V2",
-    "GeometryCHLV03_V2",
-    "GeometryCHLV95_V2",
-    "Geometry_V2",
-    "InternationalCodes_V2",
-    "Localisation_V2",
-    "LocalisationCH_V2",
-    "Dictionaries_V2",
-    "DictionariesCH_V2",
-    "CatalogueObjects_V2",
-    "CatalogueObjectTrees_V2",
 ]
 
 # style
